+++ public/html/home.html
@@ -16,13 +16,13 @@
 		<a href="/commands">Commands</a>
 	</div>
 	<div class="body">
-		<h3 id="malils-s-website">Malil&#39;s website</h3>
-		<p>
-			Malil is a all in one bot with lots of useful commands you can use in your server
+		<h2 id="malils-s-website">Malil&#39;s website</h2>
+		<h4>
+			Malil is a all in one bot with lots of useful commands.
 			<a
 				href="https://discord.com/oauth2/authorize?client_id=749020331187896410&amp;scope=bot&amp;permissions=117824">Invite
 				him here!</a>
-		</p>
+		</h4>
 
 		<p>Github command preview:</p>
 		<p class="Github">
